ev-ai-core/requirements.txt
aiohttp>=3.9.0
websockets>=12.0
click>=8.1.0
pyyaml>=6.0
certifi>=2024.0.0
